change: Split syntatic and instance implementation for nullable
The concerns are now better separated, and therefore easier to
understand and change.
