+++ include/absent/combinators/foreach.h
@@ -15,7 +15,10 @@ namespace rvarago::absent {
      */
     template <template <typename...> typename Nullable, typename Effect, typename A, typename... Rest>
     constexpr auto foreach(Nullable<A, Rest...> input, Effect fn) -> void {
-        bind(std::move(input), [&fn](auto value) { fn(std::move(value)); return Nullable<A, Rest...>{}; });
+        bind(std::move(input), [&fn](auto value) {
+            fn(std::move(value));
+            return Nullable<A, Rest...>{};
+        });
     }
 
 }
